Bug 268404. Allow formatter rules for labels
Currently formatter rules customization for labels is very limited. It is impossibe to force newline after
jump target definition, or not to ident label definition (as for example required by kernel formatter).
GNU formatter as well requires labels not to be "idented".
This makes impossible to use CDT autoformatter for development of projects that require jump targets formatting
according to either GNU or kernel standards. This fix allows to customize such behaviour and allows to specify whether jump target labels
should ident together with the code and whether they should contain newline.

Signed-off-by: Solganik Alexander <solganik@gmail.com>
Change-Id: I63a72df3557920f72be7bf9908fbe85e42538b40