Update patch set 3
Patch Set 3:

The commit message explains the change, but it basically is a change in terminology to move away from using 'trace' and use 'reverse debug' more.  Although those classes are not user-visible, I felt they made the code harder to understand for someone not used to it.  I hope the proposed changes help make it clearer (so easier to maintain).

The only functional change is different error messages when toggling reverse debugging fails.

Patch-set: 3
Label: Verified=0