Bug 435726 - Dont set SWT colors for foreground on active element.
When dark theme is in use outline and project explorer background
expected to be dark, regardless of globally (OS level) selected SWT
scheme. So for active elements foreground always return NULL which will
result in default colors to be displayed. For inactive element, continue
displaying them in grey, as grey is inactive color for both dark and
default theme.

Change-Id: I91b20b0327e0008a2aa01573981f217a609e3bbc
Signed-off-by: Solganik Alexander <solganik@gmail.com>