Update patch set 3
Patch Set 3:

When you make the first code change in a plugin since the last release, it is good to change the plugin version.  I indicates that something has been changed in that plugin.

The API tool will warn you, but only if the change affects API, which is not always true.  So, we do this on best effort.

The trick I use to know if the plugin version should be updated is that I look at the MANIFEST.MF line showing the version and I right-click on the editor margin and choose "show annotations".  This shows the git blame. I then look at the date at which the version was last changed and figure out if it was part of the ongoing release development (in which case I don't change it again) or if it was part of a previous release (in which case I do change it).

Patch-set: 3