Update patch set 1
Patch Set 1:

Math, do you need to depend on these plugins? Or was it just something you noticed?

If you need to depend on them, we could add them to the cdt.testing feature which is published in the testing update site. That way you can depend on them easily.

Patch-set: 1