Update patch set 3
Abandoned

Replace with review 20042. This change is incomplete.  First, we need logic to actually set the environment defined in the new tab.  Second, the tab does not makes sense in the manual remote launch since it is the user that sets up the gdbserver and remote environment, so we couldn't do anything about settings in an environments tab in the manual case.

Patch-set: 3
Status: abandoned