Update patch set 3
Patch Set 3: Code-Review-2

I have been thinking about this and coming back to it again and again. I can finally say that I am not comfortable that this is good enough. Fundamentally if someone extends GdbLaunch and GdbAdapterFactory and does not install the adapters properly their code is broken.

At the moment in the example with the GdbExtendedAdapterFactory it happens to all work OK because ISuspendTrigger is registered properly as on the the types GdbExtendedLaunch can adapt to and that type is one of the first ones to be adapter to. If an extender did not register one of the types that was used early on, their GdbExtendedAdapterFactory would never be used and their own session adapters would never be loaded. However the worst case is that the first time that the GdbExtendedAdapterFactory is accessed is not at the beginning of a session and in that case the behviour changes for undetermined reasons.

Therefore, I think that change 69923 should be accepted as it fixes the immediate problem of the current example being wrong.

Additionally, if extenders want to make sure that their class adapts to new types that come along in the future, then they should use a line like this in their extended GdbAdapterFactory constructor:

    Platform.getAdapterManager().registerAdapters(this, <their extended>GdbLaunch.class);

That will register programatically the adapter factory against all the types in GdbAdapterFactory.getAdapterList().

Patch-set: 3
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2