Stop using _ identifier
From JLS 15.27.1.

It is a compile-time error if a lambda parameter has the name _ (that
is, a single underscore character).

The use of the variable name _ in any context is discouraged. Future
versions of the Java programming language may reserve this name as a
keyword and/or give it special semantics.

Change-Id: I6f357dcc8f1eea933c6fc3afb474982e6d6210fe
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>