Update patch set 2
Patch Set 2: Code-Review+1

For me at W10 the bug also occurs. If your changes are applied, the issue seems to be resolved. In narrowed or non narrowed mode everything seems to work. Scrolling position is preserved too, as desired. Preventing the the "invokeLater" - runnable from being executed seems like a clean solution to me. The only thing that I would think about is: Should the scroll position of the Disassembly View be preserved, when I change threads, or is it good to reset it to current PC position like it's done now. I suggest this, because if I pre-load many instructions by scrolling down in the view they are the pre-loaded stuff is lost after a thread change.

Patch-set: 2
Reviewer: Gerrit User 76177 <76177@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1