Bug 491984 - Replace .equals("") with .isEmpty()
In many cases a String's empty status is tested with `.equals("")`. However,
Java 1.6 added `.isEmpty()` which can be more efficient since it compares the
internal length parameter only for testing. Replace code using the `.isEmpty()`
variant instead.

Some tests for `"".equals(expr)` can be replaced with `expr.isEmpty()` where
it is already known that the `expr` is not null; however, these have to be
reviewed on a case-by-case basis.

Change-Id: I3c6af4d8b7638e757435914ac76cb3a67899a5fd
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>