Update patch set 1
Patch Set 1: Code-Review-2

(7 comments)

This looks good. It works as expected. There are a number of minor issues highlighted in the comments. The most significant, but still minor, is the error message UI newline error when the symbol is not present.

However, there is a different issue that is not captured here. When the standalone application was written, these classes (NewExecutableDialog/Info) were copied/pasted. As a result, even though you have added this functionality to the full CDT case, the standalone does not benefit. 

I am not sure what the best suggestion is in that case, either to fix the divergent code, or apply these same changes to the divergent code.

Patch-set: 1
Label: Code-Review=-2