Update patch set 1
Patch Set 1: Verified+1

> Build Unstable
 > 
 > https://hudson.eclipse.org/cdt/job/cdt-verify/5079/ : UNSTABLE
Annoyingly Build 5079 failed with message:

"Timeout after: 20000 ms.: The shell 'Debug New Executable' did not activate"

but that was referring to the "other" debug new executable dialog. The re-run in build 5101 was successful.

There are not AFAIK any automated tests for this feature, I manually verified it.

Patch-set: 1
Label: Verified=+1