Update patch set 1
Patch Set 1:

(7 comments)

Thanks, I have fixed most of the issues. 
Regarding the standalone case, I was aware of it when I started working on this feature and discussed it with Marc K. We decided not to change the standalone part but now I am questioning that decision. I think we should apply these changes to the standalone part. I'll do it tomorrow.

Patch-set: 1
Label: Verified=0