Update patch set 5
Patch Set 5:

I found the issue.  The performOK doesn't run the apply logic unless the special file: apply_mode is created in the CUIPlugin state location area.  I didn't want to do this so I decided to override performOK in the ToolSettingsTab and do the logic that ultimately sets the isIndexerAffected variable.  I also had to move the calculation of rebuildIndex down in the AbstractPage performSave method so that sendok() could set it.

Patch-set: 5
Label: Verified=0