Update patch set 3
Patch Set 3: Code-Review-2

(1 comment)

The change as is does fix the problem, but as mentioned in the line comment it is duplicating work.

I certainly can't figure out if the original code was doing anything important that is not handled by the ResourceChangeHandler code. If it is not doing anything else important, I would vote for simply deleting org.eclipse.cdt.internal.core.model.PathEntryManager.updatePathEntryFromDeleteSource(ISourceRoot)

Patch-set: 3
Label: Code-Review=-2