Update patch set 1
Patch Set 1:

> Do we know if this is an intentional change of behavior in GTK3? Or
 > it could be a bug in SWT

That is next on my list. At the moment all I can see is that GTK2 and GTK3 behave differently. I was planning on raising an SWT bug one way or the other on this because it certainly isn't documented. I am using this as a GTK3 learning experience, especially in light of all the recent discussions on GTK3.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0