Update patch set 4
Patch Set 4:

> Build Failed
 > 
 > https://hudson.eclipse.org/cdt/job/cdt-verify/5150/ : FAILURE

Hi Stefan, your builds are failing because the branch is out of date. You need to rebase your changes on to the tip of the master branch and resubmit. 

You can do that from the gerrit interface (press the Rebase button) or from your egit/git environment.

Patch-set: 4