Update patch set 3
Patch Set 3:

All right, I've finally finished going over the patch, so that should be all the comments! I know there are a lot of comments, but a lot of them are trivial to address, and a lot of others are just me asking questions because I don't fully understand the code yet.

Next, I will install a version of CDT with this patch applied, and open some large codebases to see how it handles them.

Patch-set: 3
Label: Verified=0