Sign the stand-alone debugger on Mac
The mac signing has to happen between materialize-products and
archive-products goals. Because we only want to do it when the
production profile is enabled (eclipse.org server), we have to duplicate
some xml to fit the macsigner between the two. I did extract some
variables in order to make the duplication less error prone.

Change-Id: I523c287c0e1dff15026e1007fa2831af6d14003a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>