Update patch set 2
Patch Set 2: Code-Review-1

Hi Alvaro, thanks for tracking down this issue. At the moment I would prefer this is not submitted, as although it passed the tests on the build machine, tests in SourceLookupTest (specifically sourceSubstituteChanges and sourceSubstituteAdded) now fail intermittently when I run the tests. This is because the operation has stopped being atomic by changing it from a Query to a DsfRunnable.

That said, I am aware their is a tradeoff between locking the entire GUI (VERY BAD!) and a corner case of source lookup changes not working. I will take a of additional time to investigate and understand the alternatives a bit better, including seeing if I can extract the issues as additional tests.

Patch-set: 2