Bug 494650: prevent UI hang multiple launches terminated
This prevents a deadlock where two different Executor threads are both
listening to changes on the same launch configuration (e.g. when the
same launch configuration is launched twice).

See Bug 494650 for more details and information on the master branch
which has a fuller, more involved fix.

This change is a continuation of:
  commit 628389071558c43c52b666995e1eaa5c4aa67a8f
  Bug 472765: Use gdb's "set substitute-path from to"

Change-Id: I7615e34a949d7766c178025923cdab920250a591
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>