Bug 495044 - Fix launching for msys2
Not sure how this was working my desktop machine, but in my home
environment, I ran across an issue where the macros weren't resolved
in the ICdtVariables used for the launch. They need to be resolved.

I think took a look at the debug launch because it was also failing.
As I found out, the GdbLaunchDelegate wasn't using the new GDBLaunch
getVersion routine and was using the deprecated method in launch utils.

Change-Id: I63076833fc444c6df7f758e08df67c1c6f9d9f72