Bug 495623: Use new flags in =record-change from upcoming GDB release
With the next GDB (probably named 7.12), we are provided with the type
of record that was started.
  =record-started,thread-group="i1",method="btrace",format="bts"
  =record-started,thread-group="i1",method="btrace",format="pt"
  =record-started,thread-group="i1",method="full"

We make use of this information in GDBRunControl_7_10 to avoid sending
an 'info record' command.

Note that because MI adds fields in a backwards-compatible way, it was
much simpler to modify GDBRunControl_7_10 instead of creating a new
GDBRunControl_7_12 (whose name may even be wrong if GDB's next version
is not actually 7.12).

Change-Id: I869b3c0ff648d5e38081e06a7063613bbb9266dc