Update patch set 1
Patch Set 1:

(1 comment)

Thanks for writing a patch!

Could you also write a testcase for this, please? It can be added to AST2Tests.java. You can use, as guidance, another test whose body is

  parseAndCheckBindings(getAboveComment(), C);

The code to test should be the smallest piece of code that can reproduce the problem. I was able to reduce it to the following:

  struct MyStruct {
      char* str;
  };
  void foo() {
      const struct MyStruct a, b;
      (0 ? a : b).str;
  }

Patch-set: 1
Label: Code-Review=0