Bug 480271 - Update the list of features CDT claims to support in __has_feature()
__has_feature() is a clang extension for feature testing.

Additions include features that we support that weren't added previously
because they weren't listed in clang's documentation (such as cxx_atomic),
and features for which we've since gained supprot (such as
cxx_variable_templates).

The addition of cxx_atomic is particularly important because it's required
for CDT to parse libc++'s <atomic> header correctly.

Change-Id: I21a0472bb3983cf70f5b7b734827ab1111370690