Update patch set 4
Patch Set 4:

> https://git.eclipse.org/r/#/c/74999/4/core/org.eclipse.cdt.ui.tests/.settings/org.eclipse.jdt.core.prefs
 > seems to indicate that org.eclipse.jdt.core.compiler.codegen.inlineJsrBytecode
 > has to be flipped to enabled when switching from 1.7 to 1.8.

If I use "Update the classpath settings" on that plugin while keeping Java 1.7, the inlineJsrBytecode still changes.  So I don't think this is related to java 1.8.  Grepping for that setting, I see that every plugin has it enabled except cdt.ui.tests.  I haven't looked at what that setting is for, but I don't think it should be part of the script.

Patch-set: 4