Update patch set 6
Patch Set 6: -Code-Review

This turned out to be much more complicated than I expected.
Finding the right changes, without causing unwanted ones was not trivial.

I added a check at the end of the script that greps for any remaining
strings to help find anything that was missed.

I still think this is worth committing because making the changes next time
will still be tedious, so having a script will help.  But the person running the
script will need to pay close attention to the result.

Patch-set: 6