Bug 303808: Add dedicated debugger console view
Splitting out the GDB console into its own view "Debugger Console".

The goal of this patch is to allow the user to easily keep the full GDB
console in focus, without having to pin it, as the pin requirement was
not very user-friendly.  Furthermore, the user can also use the GDB
console while looking at the output of the program being debugged,
which couldn't not be done without a dedicated gdb console view.

This patch also resolves two issues we had with re-using the platform
console view, which were:
- pin didn't work
- clone didn't work
With this new Debugger console view, there is no pin and no clone.


Change-Id: Ia19132704a2f6618f35ffe47ebb4b8f0028dc9ab