Bug 497166: Support the user using the 'run' command in the gdb console
Bug 497167: Support creating a new process directly from gdb console

INCOMPLETE: 
1- Handling of attach to process
     Currently, there is a console created for processes that we attach
too.  This is wrong.
     Currently, if terminating GDB by selecting it in the DV and
pressing terminate, will not kill the console and leave the launch
hanging when dealing with an attached process
     If the user uses 'attach' from the gdb console, I don't know of a
way to differentitate it from the 'run' case, which means I can't tell
CDT not to create a console.  We could fix it when the attach is
triggered from eclipse at least, but we are looking to see if we can get
a complete solution from GDB.

2- Missing JUnit tests

Change-Id: I395b402e297a2043af8fce33df163eddef9e6c7a