Bug 497206: Remote-attach fails attach if binary not specified in launch
The core exception we used to throw when the program patch was not
present is necessary for GDBBackend#getProgramPath() to set the path to
an empty value instead of returning null.

Although we could have made GdbLaunch#getProgramPath return and empty
string to fix this problem, I thought we should play it safe and behave
like we used to in case something else needed that exception thrown.

Change-Id: I4684226c731aedef50bdeb37accdf2a2feb818b5