Deprecate model from IGenerator API.
Not all generators need a map to store information used at generate
time. In fact, this probably should be discouraged. We only had it
because Freemarker uses one. But I just wrote a generator that doesn't
use Freemaker and the model map got in the way.

Also adds a post process API so we can do other things in the Wizard
after the generation is done. Especially useful for UI things like
adding generated stuff to Working Sets.

Change-Id: Icd553fd8f6087bd342fca4aec88fb2a5c2d5fa4a