Bug 497592 - Wait for request to GDB for inferior name to be completed
GDBProcesses_7_0 would fetch the name of a process as soon as it learned
of the process through the =thread-group-started event; however, if
there was a call to IProcesses.getExecutionData() before the name was
received, the service would return the wrong name.

This commit fetches the name when IProcesses.getExecutionData() and uses
a CommandCache to do it.  That way, if a call to
IProcesses.getExecutionData() is made before the name is received, the
request will be cached until the name is available.

Furthermore, the cache allows to handle the case where the target is
unavailable, which can happen in all-stop mode and we disabled stopping
at main.

Change-Id: Ia75d3c677e189f87e9ec2864c744a870a0a18407