Update patch set 2
Patch Set 2: Code-Review+2 Verified+1

Actually, the unbrella bug is 303808, which is to support the full GDB console.  Bug 497167 is specifically about supporting an inferior console when starting a process from the GDB console, which this commit is one part of.  I think this commit is very specific, which is why I did it separatly, but since it does not provide any new funcationality yet, I don't think it deserves a bug of its own.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+2