Bug 494650: make method atomic without using a Query
Make GdbSourceLookupParticipant.sourceContainersChangedOnDispatchThread
atomic without requiring the calling thread to wrap the call in a Query.

This prevents a deadlock where two different Executor threads are both
listening to changes on the same launch configuration (e.g. when the
same launch configuration is launched twice).

See Bug 494650 for more details.

This change is a continuation of:
  commit 628389071558c43c52b666995e1eaa5c4aa67a8f
  Bug 472765: Use gdb's "set substitute-path from to"

Change-Id: I3e3faa7a079db42a709668b45e3ec5b3d473a86d
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>