Update patch set 8
Patch Set 8:

Another note about the new 'twoLaunchesTerminate' test: if I run in as o.e.ui.ide.workbench but I uncheck the "Run in UI thread" in the 'Test' tab, the test starts passing again.

I guess it makes sense because the deadlock we are trying to reproduce is triggered by actions taken on the UI thread, so we need to run the test on the UI thread.

Patch-set: 8