NOT FOR SUBMISSION: Bug 498882: This is the test for bug 494650
I am trying to reproduce the test failing properly. As it is a race
condition type failure it has proven rather difficult.

Note that this commit is based on the CDT code prior to the fix
for 494650 to verify that the test indeed fails.

Because the test is testing a full lock up, I am expecting the whole
test run to lock up and all tests to timeout once this one runs.

Change-Id: I4e50acde1654995efcf0f723d6552b68af177503
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>