Update patch set 1
Patch Set 1:

Nice to see it fail :-)
We have a 5 minute timeout for each test.  However, it seems that after the first failure, all other tests fail, probably because the whole think is locked up.  I wonder if we could cleanup nicely after this test to allow the others to run and avoid the full build timing out and not getting any test result.

Patch-set: 1
Label: Verified=0