Allow GdbLaunch#getGDBVersion() to be called more than once.
Originally, we used LaunchUtils.getGDBVersion() to fetch the GDB
version.  Because that method was used for any debug session, we didn't
cache the result of it.

Now that we moved the version handling to the GdbLaunch class, and that
this class is unique per session, we can cache the result of the
getGDBVersion(), allowing it to be called more than once.

Change-Id: I1a396134ca5c609224f8abb7b70d1e0866810497