Update patch set 5
Patch Set 5:

I have started looking at the implementation and it's looking pretty good.

The main question I have so far is about the scope of this new synchronizer service,   the service seems to be using MI interfaces, so shouldn't this new API and implementation be at the MI level?
 e.g. if the LLDB MI module supports this it can benefit from the implementation as well.

Patch-set: 5