Update patch set 5
Patch Set 5:

> The main question I have so far is about the scope of this new
 > synchronizer service,   the service seems to be using MI
 > interfaces, so shouldn't this new API and implementation be at the
 > MI level?

It is not bad to put it in the MI package, but for information purpose I wanted to mention we have stopped making that distinction.  We pretty much implement everything in GDB now.  Either that or directly in DSF.  But we don't pay attention to what is pure MI and what is specific to GDB because we don't have any other debugger using MI (see below).

 > e.g. if the LLDB MI module supports this it can benefit from the
 > implementation as well.

It turns out that LLDB MI pretends to be GDB.  So it uses GDB services and is not limited to the MI version of the service.  In fact, lldb-mi -version outputs the name 'gdb <version>' specifically to allow to use Eclipse without modification.

Patch-set: 5
Label: Verified=0