Update patch set 25
Patch Set 25: Code-Review+1

I cleaned up the patch and Patchset 25 is the result.
I'm pretty happy with this version.
Note that it still provides the full GDB console as part of the Console view.
There is a different patch in review 77986 that spits it out into its own view.
I preferred to leave the two separate so we have both options recorded in the the history.
With this in mind, one must understand that cloning the console view is not supported with the full GDB console of the current patch.  There was no point in fixing this since the plan is that the dedicated GDB console view will not be cloneable.

Patch-set: 25
Label: Code-Review=+1