Update patch set 25
Patch Set 25: Verified-1

There is one issue I am aware of in this patch that still needs to be investigated:

with the new console, when pressing terminate on the gdb node, the process node takes a second or two to disappear.  This is not the case using the old console. I believe this is because with the new console, we don't use GDBBackendCLIProcess which did some special things.  I will be looking it to this now.

Patch-set: 25
Label: Verified=-1