Bug 499230 - Duplicated "-thread-select" and "-stack-select-frame"
Most GDB MI commands support the "--thread" and "--frame" options, that
tells GDB on which thread/stack frame the command is meant to act-on.
For those that don't, CDT has a mechanism to first have GDB switch
thread (MIThreadSelect) or frame (MIStackSelectFrame), before issuing
the command that will act on that entity. Ironically, those two commands
are marked as not supporting those parameters, and so are sometimes sent
in duplicate, when a single command including the thread or frame would
be possible.

Change-Id: I6e7fc4e873abe36ad024c7131efda7c0aa89dc44