Update patch set 26
Patch Set 26:

Patchset 27 uses inheritance for the new GDBBackendProcessWithoutIO so as to avoid clode duplication.  

One the one hand, it re-uses a lot of code which makes the new class very small; on the other hand, it inherits from a class that was meant to handle the CLI commands from the user, but disables that function.  

I'm not sure if this inheritance approach is too confusing conceptually.  If it is found to be, patchset 26 can be used instead.

Patch-set: 26