Update patch set 2
Patch Set 2: Code-Review-2 Verified-1

The script as of patchset 2 will change the version of plugins in their pom.xml file,  if they happen to have the same version as what we are mathing.

We can easily notice this when the build fails because the MANIFEST.MF of that plugin won't match.  In that case, we need to revert part of the pom.xml.

Making the script smart enough to handle this is non-trivial because the pom.xml file of feature must update the version, but not the pom.xml file of plugins, so I can't simply look for pom.xml.

I'm going to leave this alone for now.
There was talk of going to pom-less builds, which could simplify this anyway.

Patch-set: 2
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=-1
Label: Code-Review=-2