Bug 497166: Support the user using the 'run' command in the gdb console
Still some issues:
1- attaching to a process probably still creates a console by mistake
2- This is a racy solution.  If the user were to source a GDB script
that quickly restarted the same process, we'd get multiple
=thread-group-started events so quickly that we would not have time to
set a new TTY for each one.  We'd end up using the old TTY and things
would crash as they used to.  It is still an improvement from before,
but maybe we can do better by re-using the same TTY.

Change-Id: I61492e51407b40fae70c6331ae55fc82713d4c20