Bug 497166: Support the user using the 'run' command in the gdb console
Still some issues:
- when we remove an exited process in eclipse, we don't remove it from
gdb.  This was not a problem before, but with the console, a user could
restart that process from GDB, but CDT would no longer create a console
for it because it was removed from the launch.  It makes more sense to
remove the inferior from GDB when it is removed from the debug session

Change-Id: I395b402e297a2043af8fce33df163eddef9e6c7a