Update patch set 2
Patch Set 2: Code-Review+2 Verified+1

I think a lot of the issues found are to be fixed in lldb-mi and not on the CDT side. I will work on fixing those. For the CDT 9.1 release, I would like to add a version check so that users will not try to use it with old versions known to fail. But I think it makes sense to be done in a separate patch and this can be done after feature freeze. I will also try to find work arounds to the other issues mentioned but it might not be possible on the CDT side.

So I'd like to go forward with this patch as it is a good start for getting the basics working and it will now be easier for people to try it and to iterate on the feature.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+2