bug 498782 - [debug] override sticky policy for starting debug session
This patch overrides the sticky selection policy for the container
started dm event. This means that a newly started debug session will get
focus in the debug view even if a stack frame from an ongoing session is
already selected. 

This harmonizes the behavior with that of the dedicated gdb console,
that switches console to follow a new debug session.

Change-Id: Ifcbffe44a6d7686c2c0daf05cfdcd821faac3a6c