Bug 303808: Disable full console in all-stop mode for the moment
We currently cannot interrupt execution with the new console in all-stop
mode. This means that if the user resumes execution then no new
breakpoints can be set from CDT while the target is running; also, the
interrupt button simply does not work.

This patch temporarily disables the new console in all-stop, until we
have the problem fixed.

Change-Id: I858268d469923700ae960600cb1dbb062729dc88