Update patch set 1
Patch Set 1:

Hi Vincent,

I would like to review this patch, however I won't be able to do it for another week or two (after I return from holiday).

One thing on quick inspection is I would be grateful for a test case on this code if feasible. Or if not feasible, how to reproduce the problem manually.

Jonah

Patch-set: 1
Label: Verified=0