Bug 337899 - [debug view][non-stop] Process label is not updated
In non-stop mode, when the last thread of a process is resumed, the
corresponding process node icon, in the Debug View, is updated to the
"running process" one. However the node was not being automatically
refreshed, and so still showed the previous "suspended process" icon.

This fix adds the necessary Delta to refresh the process node when a
IResumedDMEvent is received.

Change-Id: Ie7d2b6aef9ae7f5906e4b54470f74ee238e13ef5