Add a timeout multipler for DisplayHelper
DisplayHelper is used to run the event loop until a condition is met
or until a maximum timeout is reached. This timeout varies between
hundreds of milliseconds to a few seconds. When the tests are running
on a machine that is known to be quite under load (Hudson), the
timeouts in the milliseconds are too optimistics as there can be a
lot of other things running at the same time on the machine.

This change adds a multipler (default 1) that can be controlled from
the maven command line, for example:
-Dorg.eclipse.cdt.ui.testplugin.DisplayHelper.TIMEOUT_MULTIPLIER=5

Change-Id: I9c1517ac2641768e8ae0f4508bf9a008931ef805
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>