Update patch set 4
Patch Set 4: Code-Review-2 Verified-1

The new terminate button should apply to the session that matches the console that is showing. The current patch kills whatever is selected in the Debug view, which may not be the same as the console showing.

Also, I propse that this terminate should terminate the GDB process itself (meaning the whole session).  This may not be what is currently selected in the DV.  For example, if I disable the option to terminate GDB when the last process exists, then I select the process in the DV, then press the new Terminate button, it is the process that will be kill but the rest of the session continues to run.

Patch-set: 4
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=-1
Label: Code-Review=-2