Bug 303808: Handle resizing of full GDB console
Note that the TextCanvas used by the terminal widget that powers the
full console has a default minimum of 4 lines and 80 columns.  We could
change those if we feel that is not adequate, or if we don't want to
have such minimum values at all.  This patch leaves the minimum values
as they are by default.

Change-Id: Iad6339da1726db1102c123c97589f46ae681ffc7