Update patch set 4
Patch Set 4:

(12 comments)

I think these make the console much better.
It would have been nice if we could have updated the enablement of these commands, but I think it is not worth the effort right now; they all seem ok to be used even when they could be marked disabled; for example, if there is no text, we'd make 'clear' disabled, but pressing it again is not a problem; if there is no selection, we'd make 'copy' disabled, but leaving it enabled is not so bad (although be aware it keeps whatever was previously in the clipboard, after an empty-copy).

I have some minor comments about strings and locations of icons which I will push a new patchset for.

As a follow-up, can some of these commands be put in the basic console, so that we don't leave the basic console much worse than it was before?

Patch-set: 4
Label: Verified=0