Update patch set 5
Patch Set 5: Code-Review+2 Verified+1

There is one thing I'm not thrilled with and its that when the scroll-lock is enabled, the console does not jump to the bottom when the user types something.  I think this is how most scroll-lock work.

I realize this is a limitation in the Terminal widget though.
We can eventually fix it, or maybe we could listen for input and force it to go to the bottom?  Either way, we should get this improved.

Patch-set: 5
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1
Label: Code-Review=+2