Bug 303808: Add the DebuggerConsoleView to the debug perspective
We add the DebuggerConsoleView to the debug perspective but we do not
show it by default. This will allow us to specify the default location
of the console but not show the console unless the user actually
launches a dsf-gdb session.

The default location is specified as "stacked" to avoid taking screen
space when many users may not care about using this view.  Users that do
want to use the view a lot, will simply move it to its own space.

Change-Id: Ie7ad961193b7f72d3557022410343e325ae2af0e